Finish Max Consecutive Ones
Initial Max Consecutive Ones

When input is 1,1,0,1,1,1 and the result will be 3

When input is 1,0,1,0,1,0,1 and the result will be one

When input is 1, 1, 1, 0, 1, 0, 1 and the result also is three

When input is 0, 0, 1, 1, 1, 0, 1 and the result also is three

When input is 0, 0, 0, 0, 1, 1, 1 and the result also is three

Refactor

When input is 0 and the result will be zero

Refactor

Change the thought to current element is one, then add.
If not, reset the counter.

Input has three zero, the result also is 0

No matter input has how many zero, result alway will be zero

Refactor for loop to foreach

It beats 54.81 of submissions.

Refactor sum thoughts

Replace if condition to
counter add the current element then multiply the current element

It beats 99.04 of submissions